refactor(docs): split the landing page demo into its own folder
home-thread.tsx had grown to 1481 lines carrying the shell, the thread list,
the composer, every message part, and the memory panel. Split it into
components/pages/home/demo, one file per surface, so the next pass touches one
concern at a time. thread-specimen.tsx moves in as the demo's entry point.

MemoryInstructions and RememberToolUI were never home-page components: the docs
runtime and the toolkit mount them on every DocsRuntimeProvider surface, so they
move to components/shared/memory.tsx with their half of the test.

The sidebar now scrolls only its conversation list, with the new-thread button
and search pinned above it and an account region pinned below. Memory left the
sidebar for a view of its own, reached through a menu in the top bar, and the
composer's action row came down to the control scale the rest of the demo uses.
