+++ apps/docs/components/pages/home/demo/sidebar-context.ts
@@ -1,0 +1,7 @@
+"use client";
+
+import { createContext } from "react";
+
+export const SidebarNavigationContext = createContext<{
+  onNavigate?: (() => void) | undefined;
+}>({});
