+++ apps/docs/components/shared/memory.test.tsx
@@ -20,7 +20,7 @@ vi.mock("@/lib/memory-store", async (importOriginal) => ({
   useMemories: () => mocks.memories,
 }));
 
-import { MemoryInstructions, RememberToolUI, SidebarMemory } from "./memory";
+import { MemoryInstructions, RememberToolUI } from "./memory";
 
 afterEach(() => {
   cleanup();
@@ -49,35 +49,6 @@ describe("RememberToolUI", () => {
     );
 
     expect(onForget).toHaveBeenCalledWith("memory-1");
-  });
-});
-
-describe("SidebarMemory", () => {
-  it("renders nothing when there are no memories", () => {
-    const { container } = render(<SidebarMemory memories={[]} />);
-
-    expect(container.firstChild).toBeNull();
-  });
-
-  it("lists memories and clears them", () => {
-    const memories = [
-      {
-        id: "memory-1",
-        text: "They prefer TypeScript examples.",
-        createdAt: 1,
-      },
-    ];
-    let clear = () => {};
-    const view = render(
-      <SidebarMemory memories={memories} onClear={() => clear()} />,
-    );
-    clear = () => view.rerender(<SidebarMemory memories={[]} />);
-
-    fireEvent.click(screen.getByRole("button", { name: "Memory (1)" }));
-    expect(screen.getByText("They prefer TypeScript examples.")).toBeTruthy();
-    fireEvent.click(screen.getByRole("button", { name: "Forget all" }));
-
-    expect(view.container.firstChild).toBeNull();
   });
 });
